backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
pydantic==1.8.2
easyocr
pillow==8.3.1
spacy==3.1.2
pdf2image==1.16.0
python-docx==0.8.11
python-magic-bin